+++ iivs/dhm/data/intensity/__init__.py
@@ -12,11 +12,14 @@ __all__ = (
     "IntensityTifList",
     "IntensityTxtFolder",
     "IntensityTxtList",
+    "convert_intensity_folder",
     "load_intensity_bin",
     "load_intensity_txt",
     "read_intensity_bin_header",
     "read_intensity_txt_header",
     "save_intensity_bin",
+    "save_intensity_npy",
+    "save_intensity_txt",
 )
 
 from iivs.dhm.data.intensity.base import (
@@ -32,11 +35,13 @@ from iivs.dhm.data.intensity.bin import (
     read_intensity_bin_header,
     save_intensity_bin,
 )
-from iivs.dhm.data.intensity.npy import IntensityNpyFolder
+from iivs.dhm.data.intensity.convert import convert_intensity_folder
+from iivs.dhm.data.intensity.npy import IntensityNpyFolder, save_intensity_npy
 from iivs.dhm.data.intensity.tif import IntensityTifFolder, IntensityTifList
 from iivs.dhm.data.intensity.txt import (
     IntensityTxtFolder,
     IntensityTxtList,
     load_intensity_txt,
     read_intensity_txt_header,
+    save_intensity_txt,
 )