+++ iivs/dhm/data/phase/__init__.py
@@ -14,6 +14,7 @@ __all__ = (
     "PhaseTxtFolder",
     "PhaseTxtList",
     "PhaseUnit",
+    "convert_phase_folder",
     "convert_phase_unit",
     "load_phase_bin",
     "load_phase_txt",
@@ -21,6 +22,8 @@ __all__ = (
     "read_phase_txt_header",
     "read_phbounds",
     "save_phase_bin",
+    "save_phase_npy",
+    "save_phase_txt",
     "write_phbounds",
 )
 
@@ -38,12 +41,14 @@ from iivs.dhm.data.phase.bin import (
     save_phase_bin,
 )
 from iivs.dhm.data.phase.bounds import PhaseBounds, read_phbounds, write_phbounds
+from iivs.dhm.data.phase.convert import convert_phase_folder
 from iivs.dhm.data.phase.core import PhaseUnit, convert_phase_unit
-from iivs.dhm.data.phase.npy import PhaseNpyFolder
+from iivs.dhm.data.phase.npy import PhaseNpyFolder, save_phase_npy
 from iivs.dhm.data.phase.tif import PhaseTifFolder, PhaseTifList
 from iivs.dhm.data.phase.txt import (
     PhaseTxtFolder,
     PhaseTxtList,
     load_phase_txt,
     read_phase_txt_header,
+    save_phase_txt,
 )